+++ src/jose/jws.clj
@@ -459,8 +459,9 @@
 (defn- select-jwks-key
   [source compact]
   (let [{:keys [kid alg]} (header compact)
-        keys (jwks/get-keys source (cond-> {:alg alg}
-                                     kid (assoc :kid kid)))]
+        keys (filter jwks/signature-key?
+                     (jwks/get-keys source (cond-> {:alg alg}
+                                             kid (assoc :kid kid))))]
     (cond
       (empty? keys) (throw (jose-ex :key-not-found "No matching JWK found" nil {}))
       (and (nil? kid) (< 1 (count keys))) (throw (jose-ex :ambiguous-key "Multiple matching JWKs found" nil {}))